+++ R/about.R
@@ -1,0 +1,3 @@
+.onAttach <- function(libname, pkgname) {
+  if (!capabilities('long.double')) {options(matprod = 'internal')}
+}